Added "optional action".
Fixed bug in conditional_action_handler() where if the last data set to
execute returned a fail, this function would still return a pass. Added
"optional action" as another type of "action" that can be used, but it
always returns a pass. This allows us to perform an action, but ignore
the result if the element does not exist.