Convert remaining actions to use data_set.
Created step_result() function to move that logic out of
action_handler(). Removed android keystroke handler from
action_handler() and replaced with generic keystroke handler that will
determine if android or ios and pass the keystroke data to them. That
function already existed, but was re-written to use dependency instead
of relying on user to specify platform. Moved get_location() logic out
of action_handler() and into that funtion.