Fixed zeuz_node.py crash when server goes down, added wakeup for android.
When server goes offline, the zeuz_node.py fails to login, then exits.
This stops any further runs until it is restarted manually. Added a
loop, and a test to check if the server is up. If not, it sleeps for a
minute.

We keep getting permission errors from appium when a device is asleep.
Probably android 6.0+. Added an adb call to send KEYCODE_WAKEUP and a
swipe up to get wake the device when the appium driver is created.