Re-built appium binary search function, added testing for connected android device to make errors more clear.
Instead of assuming appium is in the path - even when it is, we have
difficulty in using it on some systems - we now have pre-programmed
locations that we expect it to be. Failing that, we'll call "which" on
Linux to help find it. Failing that, then we'll log a warning and assume
it's on the path.

Updated get_devices() from adbOptions.py, so it outputs a clean list of
connected devices instead of whatever the command output is. This was
needed, so that the new function is_android_connected() can easily check
if a device is connected. This function is now called by
find_package_name() and start_appium_driver(), to make it clear to the
user if there's a problem that it's because the device is either not
connected or not authorized.