Fixed unicode bug when running desktop automation from zeuz. Fixed bug in desktop close_program() where it can't close certain programs. Added handling of element to desktop enter_text(). Fixed bug in run_cmd() where return code was not being generated or returned.
Nothing else is affected by this, but pyautogui.locateOnScreen() kept
seeing the file_name passed to it as an object. Fixed by using encode()
to convert to ascii.

close_program() couldn't close my program, where the launcher is java.
This is too generic to close, so added the "-f" to the "pkill" command.
Also, as a backup, added code that will try to find the PID, and use
"kill" on that if pkill fails. 

In doing so, needed to update run_cmd(), so that I could get the return
code. Found it saved the return code, but also needed a wait() command
to generate it. 

To make enter_text() for desktop similar to selenium, added ability to
handle an image for the input textbox as an element paramter, similar to
click_element(). enter_text() can now use an element, but if not
provided, still assumes the textbox has keyboard focus.