Merge branch 'beta' of https://github.com/riz-hossain/ZeuzPythonNode into beta-TASK-714
# Conflicts:
#	Framework/Version.txt

All logs are marked with a errorLevel as described below  :
    iLogLevel == 0:
        status = 'Debug'  # This is not displayed on the server log, just in the console #
    iLogLevel == 1:
        status = 'Passed'
    iLogLevel == 2:
        status = 'Warning'
    iLogLevel == 3:
        status = 'Error'
    iLogLevel == 4:
        status = 'Console'
    iLogLevel == 6:
        status = 'BrowserConsole'
    Any other kind of unknown error will show :
        "*** Unknown log level- Set to Warning ***"
        status = 'Warning'

All Report style is marked with report_type as described below:
             collect_errors_only_file
                report_type=0
            elif only_errors
                report_type=1
            elif debug_mode
                report_type=2
            elif local_run
                report_type=3

ReportType was created to pass by parameter.

Also all the settings is declared and collected outside of any function . So if someone changes the settings in node while running , it will not take effect till the node is restarted. Before setting was revised in each step now not it will not.